+++ static/components/ui.js
@@ -13,4 +13,5 @@ import "./u-pagepreview.js";
 import "./u-confirm.js";
 import "./u-upload.js";
 import "./u-cropper.js";
+import "./u-tree.js";
 import "./u-busy.js";